Refactor ObjectDataInterner to use pluggable deduplicators
Extract method-body-specific deduplication logic from ObjectDataInterner
into a new MethodBodyDeduplicator class behind an IObjectDataDeduplicator
interface. ObjectDataInterner now accepts a set of deduplicator instances
and delegates the deduplication passes to them.

- Add IObjectDataDeduplicator interface with DeduplicatePass method
- Move MethodInternKey, MethodInternComparer, and CanFold logic into
  MethodBodyDeduplicator
- Add virtual CanFold method on NodeFactory (returns false by default)
- Override CanFold in RyuJitNodeFactory to delegate to MethodBodyDeduplicator
